Fix todo causing certain frontend requests to silently fail.
There was also no error logged at ContentsFileService when on L56 the contentsRequesterService failed the operation. The resulting HttpResponse was probably simply inserted into the function call's arguments.

A good practice might be to separate the code into multiple instructions and log any failure. Also verification of error codes is a good idea as seen in issue #39